requirements.txt
configparser
argparse
contextlib2
scikit-learn
communityid
PyYAML